everything ok, collisions and minimap added, need to clean code

+++ src/execution/raycasting_player_angle_adapters.c
@@ -6,7 +6,7 @@
 /*   By: vducoulo <vducoulo@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 20:42:50 by vducoulo          #+#    #+#             */
-/*   Updated: 2022/11/15 20:30:58 by vducoulo         ###   ########.fr       */
+/*   Updated: 2022/11/17 16:55:06 by vducoulo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	raycasting_horizontal_looking_up(t_game *game, t_ray *ray)
 {
-	ray->y = ((game->y >> 6) << 6) - 0.0001;
+	ray->y = (((int)game->y >> 6) << 6) - 0.0001;
 	ray->x = (game->y - ray->y) * ray->a_tan + game->x;
 	ray->y_offset = -game->game_cell_size;
 	ray->x_offset = -ray->y_offset * ray->a_tan;
@@ -24,7 +24,7 @@ void	raycasting_horizontal_looking_up(t_game *game, t_ray *ray)
 
 void	raycasting_horizontal_looking_down(t_game *game, t_ray *ray)
 {
-	ray->y = ((game->y >> 6) << 6) + game->game_cell_size;
+	ray->y = (((int)game->y >> 6) << 6) + game->game_cell_size;
 	ray->x = (game->y - ray->y) * ray->a_tan + game->x;
 	ray->y_offset = game->game_cell_size;
 	ray->x_offset = -ray->y_offset * ray->a_tan;
@@ -42,7 +42,7 @@ void	raycasting_looking_straight(t_game *game, t_ray *ray)
 
 void	raycasting_vertical_looking_right(t_game *game, t_ray *ray)
 {
-	ray->x = ((game->x >> 6) << 6) + game->game_cell_size;
+	ray->x = (((int)game->x >> 6) << 6) + game->game_cell_size;
 	ray->y = (game->x - ray->x) * ray->a_tan + game->y;
 	ray->x_offset = game->game_cell_size;
 	ray->y_offset = -ray->x_offset * ray->a_tan;
@@ -52,7 +52,7 @@ void	raycasting_vertical_looking_right(t_game *game, t_ray *ray)
 
 void	raycasting_vertical_looking_left(t_game *game, t_ray *ray)
 {
-	ray->x = ((game->x >> 6) << 6) - 0.0001;
+	ray->x = (((int)game->x >> 6) << 6) - 0.0001;
 	ray->y = (game->x - ray->x) * ray->a_tan + game->y;
 	ray->x_offset = -game->game_cell_size;
 	ray->y_offset = -ray->x_offset * ray->a_tan;
